Support live migration for config drive
Now any config drive (including vfat-based) is attached to an instance
as read-only (to prevent store data on it). This guarantee that no write
operation is preformed when live migration is going. But this leads
libvirt to skip vfat-based drives to be copied by block migration.

This fix also extends recently added feature of copying iso-based config
drives to copy any config drive via file copy operation. As the result
config drives either are forcly copied, or keep on a storage (shared
file, or rbd). Libvirt copying feature is not used for that.

Change-Id: Iacbbdd139ecb848bc99a8fb98d75e18886494e55